move logic from resolvers to service move logic from repositories to service
