+++ backend/dataall/core/environment/api/resolvers.py
@@ -17,7 +17,7 @@ from dataall.core.environment.services.environment_resource_manager import Envir
 from dataall.core.environment.services.environment_service import EnvironmentService
 from dataall.core.environment.api.enums import EnvironmentPermission
 from dataall.core.permissions.db.resource_policy_repositories import ResourcePolicy
-from dataall.core.stacks.api import stack_helper
+from dataall.core.stacks.services.stack_service import StackService
 from dataall.core.stacks.aws.cloudformation import CloudFormation
 from dataall.core.stacks.db.stack_repositories import Stack
 from dataall.core.vpc.services.vpc_service import VpcService
@@ -107,14 +107,14 @@ def create_environment(context: Context, source, input={}):
             uri=input.get('organizationUri'),
             data=input,
         )
-        Stack.create_stack(
+        StackService.create_stack(
             session=session,
             environment_uri=env.environmentUri,
             target_type='environment',
             target_uri=env.environmentUri,
             target_label=env.label,
         )
-    stack_helper.deploy_stack(targetUri=env.environmentUri)
+    StackService.deploy_stack(targetUri=env.environmentUri)
     env.userRoleInEnvironment = EnvironmentPermission.Owner.value
     return env
 
@@ -141,7 +141,7 @@ def update_environment(context: Context, source, environmentUri: str = None, inp
         )
 
         if EnvironmentResourceManager.deploy_updated_stack(session, previous_resource_prefix, environment, data=input):
-            stack_helper.deploy_stack(targetUri=environment.environmentUri)
+            StackService.deploy_stack(targetUri=environment.environmentUri)
 
     return environment
 
@@ -154,7 +154,7 @@ def invite_group(context: Context, source, input):
             data=input,
         )
 
-    stack_helper.deploy_stack(targetUri=environment.environmentUri)
+    StackService.deploy_stack(targetUri=environment.environmentUri)
 
     return environment
 
@@ -185,7 +185,7 @@ def update_group_permissions(context, source, input):
             data=input,
         )
 
-    stack_helper.deploy_stack(targetUri=environment.environmentUri)
+    StackService.deploy_stack(targetUri=environment.environmentUri)
 
     return environment
 
@@ -198,7 +198,7 @@ def remove_group(context: Context, source, environmentUri=None, groupUri=None):
             group=groupUri,
         )
 
-    stack_helper.deploy_stack(targetUri=environment.environmentUri)
+    StackService.deploy_stack(targetUri=environment.environmentUri)
 
     return environment
 
@@ -490,7 +490,7 @@ def generate_environment_access_token(context, source, environmentUri: str = Non
 
 
 def get_environment_stack(context: Context, source: Environment, **kwargs):
-    return stack_helper.get_stack_with_cfn_resources(
+    return StackService.get_stack_with_cfn_resources(
         targetUri=source.environmentUri,
         environmentUri=source.environmentUri,
     )
@@ -502,7 +502,7 @@ def delete_environment(context: Context, source, environmentUri: str = None, del
         EnvironmentService.delete_environment(session, uri=environmentUri, environment=environment)
 
     if deleteFromAWS:
-        stack_helper.delete_stack(
+        StackService.delete_stack(
             target_uri=environmentUri,
             accountid=environment.AwsAccountId,
             cdk_role_arn=environment.CDKRoleArn,
@@ -543,7 +543,7 @@ def enable_subscriptions(context: Context, source, environmentUri: str = None, i
         environment.subscriptionsConsumersTopicImported = False
         environment.subscriptionsEnabled = True
         session.commit()
-        stack_helper.deploy_stack(targetUri=environment.environmentUri)
+        StackService.deploy_stack(targetUri=environment.environmentUri)
         return True
 
 
@@ -564,7 +564,7 @@ def disable_subscriptions(context: Context, source, environmentUri: str = None):
         environment.subscriptionsProducersTopicImported = False
         environment.subscriptionsEnabled = False
         session.commit()
-        stack_helper.deploy_stack(targetUri=environment.environmentUri)
+        StackService.deploy_stack(targetUri=environment.environmentUri)
         return True
