+++ backend/dataall/core/stacks/api/queries.py
@@ -15,7 +15,6 @@ getStackLogs = gql.QueryField(
     name='getStackLogs',
     type=gql.ArrayType(gql.Ref('StackLog')),
     args=[
-        gql.Argument(name='environmentUri', type=gql.NonNullableType(gql.String)),
         gql.Argument(name='stackUri', type=gql.NonNullableType(gql.String)),
     ],
     resolver=get_stack_logs,
