+++ backend/dataall/core/stacks/api/resolvers.py
@@ -5,7 +5,7 @@ import os
 from dataall.base.api.context import Context
 from dataall.core.environment.db.environment_models import Environment
 from dataall.core.environment.services.environment_service import EnvironmentService
-from dataall.core.stacks.api import stack_helper
+from dataall.core.stacks.services.stack_service import StackService
 from dataall.core.stacks.aws.cloudformation import CloudFormation
 from dataall.core.stacks.aws.cloudwatch import CloudWatch
 from dataall.core.stacks.db.stack_models import Stack as StackModel
@@ -21,7 +21,7 @@ def get_stack(context: Context, source, environmentUri: str = None, stackUri: st
     with context.engine.scoped_session() as session:
         env: Environment = session.query(Environment).get(environmentUri)
         stack: StackModel = session.query(StackModel).get(stackUri)
-        cfn_task = stack_helper.save_describe_stack_task(session, env, stack, None)
+        cfn_task = StackService.save_describe_stack_task(session, env, stack, None)
         CloudFormation.describe_stack_resources(engine=context.engine, task=cfn_task)
         return EnvironmentService.get_stack(
             session=session,
@@ -67,33 +67,26 @@ def resolve_task_id(context, source: StackModel, **kwargs):
         return source.EcsTaskArn.split('/')[-1]
 
 
-def get_stack_logs(context: Context, source, environmentUri: str = None, stackUri: str = None):
+def get_stack_logs(context: Context, source, stackUri: str = None):
     with context.engine.scoped_session() as session:
-        stack = EnvironmentService.get_stack(session=session, uri=environmentUri, stack_uri=stackUri)
-        if not stack.EcsTaskArn:
-            raise exceptions.AWSResourceNotFound(
-                action='GET_STACK_LOGS',
-                message='Logs could not be found for this stack',
-            )
-
-        query = f"""fields @timestamp, @message, @logStream, @log as @logGroup
-                | sort @timestamp asc
-                | filter @logStream like "{stack.EcsTaskArn.split('/')[-1]}"
-                """
-        envname = os.getenv('envname', 'local')
-        results = CloudWatch.run_query(
-            query=query,
-            log_group_name=f"/{Parameter().get_parameter(env=envname, path='resourcePrefix')}/{envname}/ecs/cdkproxy",
-            days=1,
-        )
-        log.info(f'Running Logs query {query}')
-        return results
+        return StackService.get_stack_logs(session, stackUri)
 
 
 def update_stack(context: Context, source, targetUri: str = None, targetType: str = None):
+    if not targetUri:
+        raise exceptions.RequiredParameter('targetUri')
+    if not targetType:
+        raise exceptions.RequiredParameter('targetType')
     with context.engine.scoped_session() as session:
-        stack = Stack.update_stack(session=session, uri=targetUri, target_type=targetType)
-    stack_helper.deploy_stack(stack.targetUri)
+        stack = StackService.update_stack(
+            session=session,
+            targetUri=targetUri,
+            targetType=targetType,
+            username=context.username,
+            groups=context.groups,
+        )
+        StackService.deploy_stack(stack.targetUri)
+
     return stack
 
 
@@ -113,5 +106,5 @@ def update_key_value_tags(context: Context, source, input=None):
             uri=input['targetUri'],
             data=input,
         )
-        stack_helper.deploy_stack(targetUri=input['targetUri'])
+        StackService.deploy_stack(targetUri=input['targetUri'])
         return kv_tags
