+++ backend/dataall/modules/notebooks/services/notebook_service.py
@@ -14,7 +14,7 @@ from dataall.core.environment.env_permission_checker import has_group_permission
 from dataall.core.environment.services.environment_service import EnvironmentService
 from dataall.core.permissions.db.resource_policy_repositories import ResourcePolicy
 from dataall.core.permissions.permission_checker import has_resource_permission, has_tenant_permission
-from dataall.core.stacks.api import stack_helper
+from dataall.core.stacks.services.stack_service import StackService
 from dataall.core.stacks.db.keyvaluetag_repositories import KeyValueTag
 from dataall.core.stacks.db.stack_repositories import Stack
 from dataall.base.db import exceptions
@@ -138,7 +138,7 @@ class NotebookService:
                     resource_type=SagemakerNotebook.__name__,
                 )
 
-            Stack.create_stack(
+            StackService.create_stack(
                 session=session,
                 environment_uri=notebook.environmentUri,
                 target_type='notebook',
@@ -146,7 +146,7 @@ class NotebookService:
                 target_label=notebook.label,
             )
 
-        stack_helper.deploy_stack(targetUri=notebook.notebookUri)
+        StackService.deploy_stack(targetUri=notebook.notebookUri)
 
         return notebook
 
@@ -211,7 +211,7 @@ class NotebookService:
             env: Environment = EnvironmentService.get_environment_by_uri(session, notebook.environmentUri)
 
         if delete_from_aws:
-            stack_helper.delete_stack(
+            StackService.delete_stack(
                 target_uri=uri, accountid=env.AwsAccountId, cdk_role_arn=env.CDKRoleArn, region=env.region
             )
